refactor: rename source files from `.mts` to `.ts`
