+++ .gitignore
@@ -6,4 +6,4 @@
 coverage/
 node_modules/
 
-src/**/*.mjs
+src/**/*.js
