+++ action.yaml
@@ -13,4 +13,4 @@ inputs:
     required: true
 runs:
   using: node20
-  main: dist/index.mjs
+  main: dist/index.js
