+++ jest.config.json
@@ -9,9 +9,9 @@
     }
   },
   "moduleNameMapper": {
-    "^(\\.{1,2}/.*)\\.mjs$": "$1.mts"
+    "^(\\.{1,2}/.*)\\.js$": "$1"
   },
   "transform": {
-    "^.+\\.m?ts$": ["ts-jest", { "useESM": true }]
+    "^.+\\.ts$": ["ts-jest", { "useESM": true }]
   }
 }
